requirements.txt
websockets==15.0.1
httpx==0.28.1
python-dotenv==1.2.1

